chore: refactor assignee and releaser
