+++ frontend/src/components/IssueV1/logic/index.ts
@@ -2,6 +2,7 @@ export * from "./context";
 export * from "./base";
 export * from "./common";
 export * from "./assignee";
+export * from "./releaser";
 export * from "./plan";
 export * from "./review";
 export * from "./rollout";
