+++ frontend/src/components/IssueV1/logic/initialize/create.ts
@@ -40,7 +40,6 @@ import {
 import { nextUID } from "../base";
 import { sheetNameForSpec } from "../plan";
 import { createEmptyLocalSheet, getLocalSheetByName } from "../sheet";
-import { trySetDefaultAssignee } from "./assignee";
 
 export type InitialSQL = {
   sqlList?: string[];
@@ -96,8 +95,6 @@ export const createIssueSkeleton = async (query: Record<string, string>) => {
   const rollout = await previewPlan(plan, params);
   issue.rollout = rollout.name;
   issue.rolloutEntity = rollout;
-
-  await trySetDefaultAssignee(issue);
 
   const description = query.description;
   if (description) {
